Use frames instead of seconds in age field

+++ ParticleData.cs
@@ -6,7 +6,7 @@ namespace ParticleGame
     {
         public ParticleTypes.Types ParticleType { get; set; }
         public Point PreviousPosition { get; set; }
-        public float Age { get; set; }
+        public int Age { get; set; }
         public bool Awake { get; set; }
 
         public ParticleData(ParticleTypes.Types particleType, Point previousPosition)
